app.py
import os
import base64
from flask import Flask, request, jsonify
from flask_cors import CORS
from dotenv import load_dotenv
from openai import OpenAI

# --- إعدادات التطبيق الأساسية ---
app = Flask(__name__, static_folder='frontend', static_url_path='')
CORS(app) 
load_dotenv()

# --- تهيئة عميل OpenAI ---
client = OpenAI()

# =================================================================================
# --- (تم تحديث هذا الجزء بالكامل) ---
# --- قاموس البرومبتات الاحترافي والمقويات (Professional Prompts & Enhancers Dictionary) ---
# =================================================================================

PROMPT_SYSTEMS = {
    "general": """You are a meticulous technical image analyst. Your mission is to deconstruct an image into its fundamental visual components with objective precision and then enrich it with artistic terminology.

**CRITICAL INSTRUCTIONS:**
1.  **Subject Analysis:** Describe the main subject's key features: age, gender, ethnicity, hair style and color, clothing, expression, and physical build.
2.  **Action & Position:** State exactly what the subject is doing and where they are positioned within the frame.
3.  **Environment Analysis:** Detail the immediate surroundings, including objects, furniture, and background elements.
4.  **Composition & Lighting:** Describe the shot type, camera angle, and the characteristics of the lighting and dominant colors.
5.  **Enrichment:** After your initial analysis, review the following keyword palettes. Select and **seamlessly integrate the MOST RELEVANT terms** into your descriptive paragraph to enhance its technical and artistic depth.
6.  **Final Output:** Combine all observations into a single, detailed, and coherent descriptive paragraph. **DO NOT** list the keywords. Weave them naturally into the sentences.

---
**KEYWORD PALETTES FOR ENRICHMENT:**

* **🔥 1. مقويات الأسلوب السينمائي (Cinematic Enhancers):** cinematic lighting, dramatic composition, ultra-wide aspect ratio, Rembrandt lighting, chiaroscuro shadows, depth of field, backlighting / rim lighting, volumetric light beams, bokeh background, film noir style, hyper-realistic.
* **🎨 2. مقويات التفاصيل والدقة (Detail & Texture Enhancers):** 8k resolution, ultra-detailed, photorealistic, intricate textures, high contrast, finely rendered, HDR lighting, sharp focus, subtle skin pores / fabric wrinkles.
* **🧠 3. مقويات الشخصية والتعبير (Character & Emotion Boosters):** intense expression, wild eyes, manic grin, calm but deadly look, focused gaze, heroic stance, emotional tension, mysterious smile, sorrowful elegance, confident posture.
* **🧪 4. مقويات الجو والبيئة (Atmosphere Enhancers):** moody atmosphere, mysterious setting, glowing fog / vapor, enchanted forest / arcane lab, steampunk elements, magical runes / symbols, glowing particles, dystopian vibe, bioluminescent elements.
---
""",

    "midjourney": """You are a world-class AI prompt engineer for Midjourney. Your goal is to transform an image into the most effective, detailed, and clean prompt possible by combining direct analysis with a rich vocabulary of artistic keywords.

**CRITICAL INSTRUCTIONS:**
1.  Analyze the image in-depth: subject, setting, style, lighting, colors, composition.
2.  From the keyword palettes below, select the most relevant terms to enhance the prompt.
3.  Synthesize all details and selected keywords into a single, continuous string of text.
4.  **Your final output MUST be a clean, comma-separated list of keywords and descriptive phrases.**
5.  **DO NOT use numbers, bullet points, or labels.**
6.  Conclude with appropriate Midjourney parameters like `--ar` (based on image orientation) and `--v 6.0`.

---
**KEYWORD PALETTES:**

* **🔥 1. Cinematic Enhancers:** cinematic lighting, dramatic composition, Rembrandt lighting, chiaroscuro shadows, depth of field, backlighting / rim lighting, volumetric light beams, bokeh background, film noir style.
* **🎨 2. Detail & Texture Enhancers:** 8k, ultra-detailed, hyperrealistic, intricate textures, high contrast, finely rendered, HDR, sharp focus.
* **🧠 3. Character & Emotion Boosters:** intense expression, wild eyes, manic grin, calm but deadly look, focused gaze, heroic stance, emotional tension, mysterious smile, sorrowful elegance, confident posture.
* **🧪 4. Atmosphere Enhancers:** moody atmosphere, mysterious setting, glowing fog / vapor, enchanted forest / arcane lab, steampunk elements, magical runes / symbols, glowing particles, dystopian vibe, bioluminescent elements.
---
""",

    "stable-diffusion": """You are an expert prompt creator for SDXL (Stable Diffusion). Your goal is to generate a high-quality prompt that captures the essence of the image by layering descriptive phrases with powerful keywords.

**CRITICAL INSTRUCTIONS:**
1.  Start with a clear description of the main subject and action.
2.  Follow up with comma-separated keywords detailing the style, environment, and artistic medium.
3.  From the keyword palettes below, select and add the most relevant terms to maximize detail, realism, and mood.
4.  **End the prompt with the most powerful quality-boosting tags** (e.g., `masterpiece, best quality, absurdres`).
5.  The entire prompt should be a single, flowing sentence structure combined with comma-separated keywords.

---
**KEYWORD PALETTES:**

* **🔥 1. Cinematic Enhancers:** cinematic lighting, dramatic composition, Rembrandt lighting, chiaroscuro shadows, depth of field, backlighting / rim lighting, volumetric light beams, bokeh background.
* **🎨 2. Detail & Texture Enhancers:** 8k resolution, ultra-detailed, photorealistic, intricate textures, high contrast, finely rendered, HDR lighting, sharp focus, subtle skin pores.
* **🧠 3. Character & Emotion Boosters:** intense expression, wild eyes, manic grin, focused gaze, heroic stance, emotional tension, mysterious smile.
* **🧪 4. Atmosphere Enhancers:** moody atmosphere, mysterious setting, glowing fog / vapor, enchanted forest / arcane lab, steampunk elements, magical runes, glowing particles, dystopian vibe.
---
""",

    "flux": """You are a prompt engineer for the Flux model, valuing clarity and conciseness.
**CRITICAL INSTRUCTIONS:**
1. Create a straightforward but descriptive prompt.
2. Focus on a clear `subject, action, setting` structure.
3. Use simple, powerful keywords.
**Example:** `A majestic lion with a flowing mane, standing proudly on a rocky outcrop overlooking the savanna at sunset, warm golden light, photorealistic, high detail.`"""
}


# =============================================================
# --- المسارات (Routes) ---
# =============================================================

@app.route('/')
def tool_page():
    return app.send_static_file('index.html')


@app.route('/api/generate-prompt', methods=['POST'])
def generate_prompt():
    if not request.json:
        return jsonify({"error": "Invalid request. Expecting JSON."}), 400

    image_data_url = request.json.get('image_base64')
    model_type = request.json.get('model', 'general')

    if not image_data_url:
        return jsonify({"error": "No image data provided."}), 400
    
    system_prompt = PROMPT_SYSTEMS.get(model_type, PROMPT_SYSTEMS['general'])
    
    try:
        response = client.chat.completions.create(
            model="gpt-4o",
            messages=[
                {
                    "role": "system",
                    "content": system_prompt
                },
                {
                    "role": "user",
                    "content": [
                        {
                            "type": "text",
                            "text": "Analyze this image and generate a prompt based on my system instructions."
                        },
                        {
                            "type": "image_url",
                            "image_url": { "url": image_data_url }
                        }
                    ]
                }
            ],
            max_tokens=1000
        )
        description_text = response.choices[0].message.content.strip()
        return jsonify({"prompt": description_text})

    except Exception as e:
        print(f"An error occurred: {e}")
        return jsonify({"error": str(e)}), 500


# --- تشغيل التطبيق ---
if __name__ == '__main__':
    app.run(debug=True, port=5000)